fixes toml provider to write toml objects, fixes toml writing to not override old documents and preserve comments
