+++ dlt/common/configuration/providers/__init__.py
@@ -4,12 +4,13 @@ from .dictionary import DictionaryProvider
 from .toml import (
     SecretsTomlProvider,
     ConfigTomlProvider,
-    ProjectDocProvider,
+    SettingsTomlProvider,
     CONFIG_TOML,
     SECRETS_TOML,
     StringTomlProvider,
     CustomLoaderDocProvider,
 )
+from .doc import CustomLoaderDocProvider
 from .vault import SECRETS_TOML_KEY
 from .google_secrets import GoogleSecretsProvider
 from .context import ContextProvider
@@ -20,7 +21,7 @@ __all__ = [
     "DictionaryProvider",
     "SecretsTomlProvider",
     "ConfigTomlProvider",
-    "ProjectDocProvider",
+    "SettingsTomlProvider",
     "CONFIG_TOML",
     "SECRETS_TOML",
     "StringTomlProvider",
