+++ tests/common/configuration/test_toml_provider.py
@@ -16,6 +16,7 @@ from dlt.common.configuration.providers.toml import (
     CONFIG_TOML,
     BaseDocProvider,
     CustomLoaderDocProvider,
+    SettingsTomlProvider,
     SecretsTomlProvider,
     ConfigTomlProvider,
     StringTomlProvider,
@@ -246,7 +247,7 @@ def test_toml_get_key_as_section(toml_providers: ConfigProvidersContext) -> None
 def test_toml_read_exception() -> None:
     pipeline_root = "./tests/common/cases/configuration/.wrong.dlt"
     with pytest.raises(TomlProviderReadException) as py_ex:
-        ConfigTomlProvider(project_dir=pipeline_root)
+        ConfigTomlProvider(settings_dir=pipeline_root)
     assert py_ex.value.file_name == "config.toml"
 
 
@@ -288,7 +289,7 @@ def test_toml_global_config() -> None:
 
 
 def test_write_value(toml_providers: ConfigProvidersContext) -> None:
-    provider: BaseDocProvider
+    provider: SettingsTomlProvider
     for provider in toml_providers.providers:  # type: ignore[assignment]
         if not provider.is_writable:
             continue
@@ -351,9 +352,10 @@ def test_write_value(toml_providers: ConfigProvidersContext) -> None:
             "dict_test.deep_dict.embed.inner_2",
         )
         # write a dict over non dict
-        provider.set_value("deep_list", test_d1, None, "deep", "deep", "deep")
+        ovr_dict = {"ovr": 1, "ocr": {"ovr": 2}}
+        provider.set_value("deep_list", ovr_dict, None, "deep", "deep", "deep")
         assert provider.get_value("deep_list", TAny, None, "deep", "deep", "deep") == (
-            test_d1,
+            ovr_dict,
             "deep.deep.deep.deep_list",
         )
         # merge dicts
@@ -368,7 +370,8 @@ def test_write_value(toml_providers: ConfigProvidersContext) -> None:
             test_m_d1_d2,
             "dict_test.deep_dict",
         )
-        # print(provider.get_value("deep_dict", Any, None, "dict_test"))
+        # compare toml and doc repr
+        assert provider._config_doc == provider._config_toml.unwrap()
 
         # write configuration
         pool = PoolRunnerConfiguration(pool_type="none", workers=10)
@@ -403,7 +406,7 @@ def test_set_spec_value(toml_providers: ConfigProvidersContext) -> None:
 
 
 def test_set_fragment(toml_providers: ConfigProvidersContext) -> None:
-    provider: BaseDocProvider
+    provider: SettingsTomlProvider
     for provider in toml_providers.providers:  # type: ignore[assignment]
         if not isinstance(provider, BaseDocProvider):
             continue
